Make schedule methods available to all users
* Move schedule-related methods from `student_methods` model to `person_methods` model to make them global of all users (staff and students alike).
