+++ src/models/methods/person_methods.js
@@ -1,7 +1,7 @@
 // Methods for person/all user account types
 
 const { v4: uuid } = require('uuid');
-const { enums, privileges } = require('../base');
+const { enums, privileges, immutables } = require('../base');
 
 const { statuses } = enums.students;
 
@@ -101,6 +101,111 @@ async function changePassword(newPassword) {
   return this.save();
 }
 
+/**
+ * Class method for retrieving schedules for a student by date. Result is sorted by `time` field.
+ * @param {Date} startDate Lower bound of date range.
+ * @param {Date} endDate Upper bound of date range.
+ * @returns {Promise.<mongoose.Model[]>}
+ */
+// async function getSchedules(startDate = '1900-01-01', endDate = Date.now()) {
+async function getSchedules(startDate = '1900-01-01', endDate = '2999-12-31') {
+  if (new Date(startDate).toString() === 'Invalid Date') return { error: 'ValueError: Invalid startDate' };
+  if (new Date(endDate).toString() === 'Invalid Date') return { error: 'ValueError: Invalid endDate' };
+
+  // Retrieve schedules from database
+  return mongoose.model('Schedule').find({
+    createdBy: this.id,
+    time: { $gte: startDate, $lte: endDate },
+  }).sort({ time: 1 }).select({ ...privateAttr.all, createdBy: 0 });
+}
+
+/**
+ * Class method for retrieving schedules for a student by date. Result is a parsed object.
+ * @param {Date} startDate Lower bound of date range.
+ * @param {Date} endDate Upper bound of date range.
+ * @returns {Promise.<Object>}
+ */
+async function getParsedSchedules(startDate = '1900-01-01', endDate = '2999-12-31') {
+  // Retrieve schedules from database
+  const schedules = await this.getSchedules(startDate, endDate);
+  if (schedules.error) return { error: schedules.error };
+  if (!schedules || !schedules.length) return {};
+
+  // Parse and return schedules data
+  return schedules.reduce((results, schedule) => {
+    const [day, month, year] = schedule.time
+      .toLocaleDateString('en-gb', { day: 'numeric', month: 'long', year: 'numeric' }).split(' ');
+    if (!results[`${month}-${year}`]) results[`${month}-${year}`] = {}; // Add slot for month and year
+    if (!results[`${month}-${year}`][day]) results[`${month}-${year}`][day] = []; // Add slot for day
+    results[`${month}-${year}`][day].push(schedule);
+    return results;
+  }, {});
+}
+
+/**
+ * Class method for creating a new schedule for a student
+ * @param {Object} attributes Attributes to be assigned to the new schedule.
+ * @returns {Promise.<mongoose.Model>}
+ */
+async function createSchedule(attributes) {
+  if (!attributes || typeof attributes !== 'object') return { error: 'ValueError: Invalid attributes' };
+
+  // Prevent setting user-immutable attributes on this schedule
+  for (const key of immutables.Project) { delete attributes[key]; }
+
+  attributes.createdBy = this.id; // Permanently link the new schedule to this student
+
+  return mongoose.model('Schedule')(attributes).save();
+}
+
+/**
+ * Class method for updating an existing student's schedule.
+ * @param {ObjectId} id ID of schedule to be updated.
+ * @param {Object} attributes New attributes to be assigned to the schedule.
+ * @returns {Promise.<mongoose.Model>}
+ */
+async function updateSchedule(id, attributes) {
+  if (!ObjectId.isValid(id)) return { error: 'ValueError: Invalid id' };
+  if (!attributes || typeof attributes !== 'object') return { error: 'ValueError: Invalid attributes' };
+
+  // Retrieve schedule from database
+  const schedule = await mongoose.model('Schedule').findById(id).exec();
+  if (!schedule) return { error: `ValueError: Schedule with id=${id} not found` };
+
+  // Check if schedule is owned by this student
+  if (String(schedule.createdBy) !== this.id) return { error: 'Access denied' };
+
+  // Prevent updating user-immutable attributes on this schedule
+  for (const key of immutables.Project) { delete attributes[key]; }
+
+  // Apply given attributes updates to schedule
+  for (const [key, value] of Object.entries(attributes)) {
+    schedule[key] = value;
+  }
+
+  return schedule.save();
+}
+
+/**
+ * Class method for deleting a student's schedule.
+ * @param {ObjectId} id ID of schedule to be deleted.
+ * @returns {Promise.<Boolean>}
+ */
+async function deleteSchedule(id) {
+  if (!ObjectId.isValid(id)) return { error: 'ValueError: Invalid id' };
+
+  // Retrieve schedule from database
+  const schedule = await mongoose.model('Schedule').findById(id).exec();
+  if (!schedule) return { error: `ValueError: Schedule with id=${id} not found` };
+
+  // Check if schedule is owned by this student
+  if (String(schedule.createdBy) !== this.id) return { error: 'Access denied' };
+
+  // Delete schedule from database and return result
+  if ((await schedule.deleteOne()).deletedCount >= 1) return true;
+  return false;
+}
+
 module.exports = {
   validatePerson,
   getFullName,
@@ -109,4 +214,9 @@ module.exports = {
   validateOTP,
   resetPassword,
   changePassword,
+  getSchedules,
+  getParsedSchedules,
+  createSchedule,
+  updateSchedule,
+  deleteSchedule,
 };
